imageProject/requirements.txt
aiohttp==3.7.4.post0
anytree==2.8.0
apipkg==2.1.1
apparmor==3.0.4
appdirs==1.4.4
application-utility==1.3.2
async-timeout==3.0.1
attrs==21.4.0
Beaker==1.11.0
btrfsutil==5.17
build==0.7.0
CacheControl==0.12.6
cachy==0.3.0
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
cleo==0.8.1
click==8.1.3
clikit==0.6.2
colorama==0.4.4
contextlib2==0.6.0.post1
coverage==6.1.1
crashtest==0.3.1
cryptography==37.0.0
cs50==9.0.0
cupshelpers==1.0
discord==1.7.3
discord.py==1.7.3
distlib==0.3.4
distro==1.7.0
dnspython==2.2.1
docopt==0.6.2
entrypoints==0.4
evdev==1.5.0
file-magic==0.4.0
filelock==3.5.0
first==2.0.2
Flask==2.1.2
greenlet==1.1.2
hidapi==0.11.2
html5lib==1.1
idna==3.3
importlib-metadata==4.8.1
iniconfig==1.1.1
installer==0.5.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==3.2.0
keyring==23.4.0
keyutils==0.6
kiss-headers==2.3.1
lark-parser==0.12.0
lensfun==0.3.3
LibAppArmor==3.0.4
libtorrent==2.0.6
lit==13.0.1.dev0
lockfile==0.12.2
Logbook==1.5.3
louis==3.21.0
lxml==4.8.0
Mako==1.1.6
mallard-ducktype==1.0.2
manjaro-sdk==0.1
Markdown==3.3.7
MarkupSafe==2.1.1
meson==0.62.1
more-itertools==8.12.0
msgpack==1.0.3
multidict==6.0.2
netsnmp-python==1.0a1
npyscreen==4.10.5
numpy==1.22.3
ordered-set==4.0.2
packaging==21.3
pacman-mirrors==4.23.2
pastel==0.2.1
pep517==0.12.0
pexpect==4.8.0
Pillow==9.1.0
pkginfo==1.8.2
platformdirs==2.5.0
pluggy==1.0.0
ply==3.11
poetry==1.1.13
poetry-core==1.0.8
progress==1.6
ptyprocess==0.7.0
pwquality==1.4.4
py==1.11.0
PyAudio==0.2.11
pycairo==1.21.0
pycparser==2.21
pycups==2.0.1
pycurl==7.45.1
Pygments==2.12.0
PyGObject==3.42.1
pylev==1.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.6
PyQt5-sip==12.10.1
pyrsistent==0.18.1
PySide6==6.3.0
pysmbc==1.0.23
pytest==7.1.2
pytest-cov==2.12.1
pyxdg==0.27
PyYAML==6.0
rados==2.0.0
rbd==2.0.0
reportlab==3.6.9
requests==2.27.1
requests-toolbelt==0.9.1
resolvelib==0.5.5
retrying==1.3.3
rgw==2.0.0
rivalcfg==4.5.0
SecretStorage==3.3.1
shellingham==1.4.0
shiboken6==6.3.0
single-version==1.5.1
six==1.16.0
SQLAlchemy==1.4.36
sqlparse==0.4.2
systemd-python==234
team==1.0
termcolor==1.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.10.2
typing-extensions==4.2.0
udiskie==2.4.2
urllib3==1.26.9
virtualenv==20.11.0
webencodings==0.5.1
websockets==10.3
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0